+++ requirements.txt
@@ -1,0 +1,57 @@
+aiohttp==3.8.5
+aiosignal==1.3.1
+amqp==5.1.1
+appnope==0.1.3
+asttokens==2.4.0
+async-timeout==4.0.3
+attrs==23.1.0
+backcall==0.2.0
+billiard==4.1.0
+blinker==1.6.2
+celery==5.3.4
+certifi==2023.7.22
+charset-normalizer==3.2.0
+click==8.1.7
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.3.0
+decorator==5.1.1
+executing==1.2.0
+Flask==2.3.3
+Flask-Cors==4.0.0
+frozenlist==1.4.0
+gensim==4.3.2
+idna==3.4
+ipython==8.15.0
+itsdangerous==2.1.2
+jedi==0.19.0
+Jinja2==3.1.2
+kombu==5.3.2
+MarkupSafe==2.1.3
+matplotlib-inline==0.1.6
+multidict==6.0.4
+numpy==1.26.0
+openai==0.28.0
+parso==0.8.3
+pexpect==4.8.0
+pickleshare==0.7.5
+prompt-toolkit==3.0.39
+ptyprocess==0.7.0
+pure-eval==0.2.2
+Pygments==2.16.1
+python-dateutil==2.8.2
+python-dotenv==1.0.0
+redis==5.0.0
+requests==2.31.0
+scipy==1.11.2
+six==1.16.0
+smart-open==6.4.0
+stack-data==0.6.2
+tqdm==4.66.1
+traitlets==5.10.0
+tzdata==2023.3
+urllib3==2.0.4
+vine==5.0.0
+wcwidth==0.2.6
+Werkzeug==2.3.7
+yarl==1.9.2
